# apps/api/internal/postgres/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlc

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Session struct {
	TokenHash         []byte             `json:"token_hash"`
	UserID            string             `json:"user_id"`
	CreatedAt         pgtype.Timestamptz `json:"created_at"`
	LastSeenAt        pgtype.Timestamptz `json:"last_seen_at"`
	IdleExpiresAt     pgtype.Timestamptz `json:"idle_expires_at"`
	AbsoluteExpiresAt pgtype.Timestamptz `json:"absolute_expires_at"`
}

type User struct {
	ID           string             `json:"id"`
	Email        string             `json:"email"`
	FirstName    string             `json:"first_name"`
	LastName     string             `json:"last_name"`
	PasswordHash string             `json:"password_hash"`
	Role         string             `json:"role"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	UpdatedAt    pgtype.Timestamptz `json:"updated_at"`
}