# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Платформа подбора роботизированных решений API
  version: 0.1.0
  description: Контракт API авторизации платформы.

servers:
  - url: /api

tags:
  - name: Auth
    description: Регистрация, вход и текущая серверная сессия.

paths:
  /auth/register:
    post:
      tags:
        - Auth
      operationId: register
      summary: Зарегистрировать пользователя
      description: Создает учетную запись с ролью user и устанавливает серверную сессию.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Пользователь создан, серверная сессия установлена.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Запрос не прошел проверку.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidEmail:
                  $ref: '#/components/examples/InvalidEmail'
                invalidFirstName:
                  $ref: '#/components/examples/InvalidFirstName'
                invalidLastName:
                  $ref: '#/components/examples/InvalidLastName'
                invalidPassword:
                  $ref: '#/components/examples/InvalidPassword'
        '409':
          description: Канонический email уже занят.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: email_already_registered
                message: Пользователь с таким email уже зарегистрирован.
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/login:
    post:
      tags:
        - Auth
      operationId: login
      summary: Войти в систему
      description: Проверяет email и пароль и устанавливает серверную сессию.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Вход выполнен, серверная сессия установлена.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Запрос не прошел проверку.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Email или пароль неверны.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: invalid_credentials
                message: Неверный email или пароль.
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  /auth/me:
    get:
      tags:
        - Auth
      operationId: getCurrentUser
      summary: Получить текущего пользователя
      description: Возвращает пользователя, связанного с действующей серверной сессией.
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Текущий пользователь.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Действующая сессия отсутствует или истекла.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: unauthenticated
                message: Требуется вход в систему.
        '500':
          $ref: '#/components/responses/InternalError'

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
      description: Идентификатор серверной сессии. Cookie не содержит пароль или данные пользователя.

  headers:
    SessionCookie:
      description: Сессионная cookie с атрибутами HttpOnly, SameSite=Lax и Secure в сетевом окружении.
      schema:
        type: string

  responses:
    InternalError:
      description: Внутренняя ошибка сервера.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: internal_error
            message: Внутренняя ошибка сервера.

    RateLimited:
      description: Превышен лимит запросов.
      headers:
        Retry-After:
          description: Число секунд до следующей разрешенной попытки.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: rate_limited
            message: Слишком много запросов. Повторите попытку позже.

  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - firstName
        - lastName
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: Email. Сервер удаляет крайние пробелы и каноникализирует адрес перед проверкой уникальности.
          examples:
            - user@example.com
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          description: Имя без управляющих символов. Сервер удаляет крайние пробелы.
          examples:
            - Иван
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          description: Фамилия без управляющих символов. Сервер удаляет крайние пробелы.
          examples:
            - Петров
        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
          description: Пароль длиной от 12 до 128 символов.

    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          examples:
            - user@example.com
        password:
          type: string
          format: password
          writeOnly: true

    User:
      type: object
      additionalProperties: false
      required:
        - id
        - email
        - firstName
        - lastName
        - role
      properties:
        id:
          type: string
          description: Уникальный идентификатор учетной записи.
          examples:
            - 01JQ2Q5K3D7H5Y8C2J4M6N8P0R
        email:
          type: string
          format: email
          readOnly: true
        firstName:
          type: string
          readOnly: true
        lastName:
          type: string
          readOnly: true
        role:
          type: string
          enum:
            - user
            - admin
          readOnly: true

    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Машинно-читаемый код ошибки.
          examples:
            - invalid_email
        message:
          type: string
          description: Сообщение для отображения пользователю.
        details:
          type: object
          additionalProperties: true
          description: Дополнительные данные об ошибке, если они нужны клиенту.

  examples:
    InvalidEmail:
      value:
        code: invalid_email
        message: Укажите корректный email.
    InvalidFirstName:
      value:
        code: invalid_first_name
        message: Имя должно содержать от 1 до 100 символов.
    InvalidLastName:
      value:
        code: invalid_last_name
        message: Фамилия должна содержать от 1 до 100 символов.
    InvalidPassword:
      value:
        code: invalid_password
        message: Пароль должен содержать от 12 до 128 символов.